feat: minimum support for custom parameter types
Add support for parameter types in the same interface as cucumber-js.

Note that unlike the cucumber-js, this commit requires type definition
in the file that is loaded prior to the step definition.
For example, a file with a location and name such as
cypress/integration/00parameter_types.js should be used to define custom
parameter types.